refactor: streamline token and ID retrieval in RequireLogin component
